+++ Source/Krypton Components/Krypton.Toolkit/Palette Controls/PaletteInputControlContentStates.cs
@@ -23,6 +23,7 @@ public class PaletteInputControlContentStates : Storage,
     private Font? _font;
     private Color _color1;
     internal Padding _padding;
+    private PaletteRelativeAlign _factoryTextH = PaletteRelativeAlign.Inherit;
     internal PaletteRelativeAlign _shortTextH;
 
     #endregion
@@ -62,6 +63,11 @@ public class PaletteInputControlContentStates : Storage,
                                          (Color1.IsEmpty) &&
                                          Padding.Equals(CommonHelper.InheritPadding)
                                          && !ShouldSerializeTextH();
+
+    /// <summary>
+    /// Treats the current values as the unset designer default.
+    /// </summary>
+    public void CaptureFactoryDefaults() => _factoryTextH = _shortTextH;
 
     #endregion
 
@@ -209,7 +215,6 @@ public class PaletteInputControlContentStates : Storage,
     [Category(@"Visuals")]
     [Description(@"Relative horizontal Content text alignment\nIn order to get this into the designer.cs you must also modify another value in this area!")]
     [RefreshProperties(RefreshProperties.All)]
-    [DefaultValue(PaletteRelativeAlign.Inherit)]
     public PaletteRelativeAlign TextH
     {
         get => _shortTextH;
@@ -224,16 +229,17 @@ public class PaletteInputControlContentStates : Storage,
         }
     }
 
-    private bool ShouldSerializeTextH() => _shortTextH != PaletteRelativeAlign.Inherit;
+    private bool ShouldSerializeTextH() => _shortTextH != _factoryTextH;
 
-    private void ResetTextH() => _shortTextH = PaletteRelativeAlign.Inherit;
+    private void ResetTextH() => _shortTextH = _factoryTextH;
 
     /// <summary>
     /// Gets the actual content short text horizontal alignment value.
     /// </summary>
     /// <param name="state">Palette value should be applicable to this state.</param>
     /// <returns>RelativeAlignment value.</returns>
-    public virtual PaletteRelativeAlign GetContentShortTextH(PaletteState state) => ShouldSerializeTextH() ? _shortTextH : Inherit.GetContentShortTextH(state);
+    public virtual PaletteRelativeAlign GetContentShortTextH(PaletteState state) =>
+        _shortTextH != PaletteRelativeAlign.Inherit ? _shortTextH : Inherit.GetContentShortTextH(state);
 
     /// <summary>
     /// Gets the actual content short text vertical alignment value.